+++ composer.json
@@ -42,7 +42,7 @@
         "mockery/mockery": "^1.5.1",
         "orchestra/canvas": "^8.0",
         "phpstan/phpstan": "^1.8",
-        "phpunit/phpunit": "^9.5.10",
+        "phpunit/phpunit": "^9.5.27 || ^10.0",
         "symfony/process": "^6.1",
         "symfony/yaml": "^6.1",
         "vlucas/phpdotenv": "^5.4.1"
@@ -55,7 +55,7 @@
         "nunomaduro/collision": "Allow using Laravel style tests output and parallel testing (^6.2).",
         "orchestra/testbench-browser-kit": "Allow using legacy Laravel BrowserKit for testing (^7.0).",
         "orchestra/testbench-dusk": "Allow using Laravel Dusk for testing (^7.0).",
-        "phpunit/phpunit": "Allow using PHPUnit for testing (^9.5.10).",
+        "phpunit/phpunit": "Allow using PHPUnit for testing (^9.5.10 || ^10.0).",
         "symfony/yaml": "Required for CLI Commander (^6.1).",
         "vlucas/phpdotenv": "Required for CLI Commander (^5.4.1)."
     },
